# Optimize `escape_ascii` using a lookup table