@bugadani 2024-05-28
> but 256 bytes should be small enough that very few people will mind

I'm just driving by but maybe this change could be behind [optimize_for_size](https://github.com/rust-lang/rust/pull/125011)