@clarfonthey 2024-05-28
I actually saw the `optimize_for_size` flag but kind of forgot to even try that here. I'm not 100% sure that the old algorithm will always be smaller after optimization, but it probably will on architectures that users of `opitmize_for_size` care about.

Did a tiny bit of refactoring to share more code between the two versions, and it appeared to not affect the benchmarks in a statistically significant way.

I'll *probably* take a poke later at the contents of #125317 and "rebase" this on top of that, so that this version also includes the `MaybeUninit` improvements.