@clarfonthey 2024-10-09
Okay, this should have the reflected changes + additional tests to ensure it passes on both regular + optimize-for-size mode.

@rustbot review